deploy-platform/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
ansible-runner==2.3.4
python-dotenv==1.0.0
bcrypt==4.1.2